Moved DateInput from common to utils
